refactor: 문자열 파싱을 컨트롤러가 아닌 서비스에서 수행
