+++ src/main/java/lotto/service/LottoService.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface LottoService {
 
-    double computeProfitRate(int purchaseAmount, List<Integer> winningNumbers, int bonusNumber);
+    double computeProfitRate(String purchaseAmount, String winningNumbers, String bonusNumber);
 }
